requirements.txt
# Core data science
numpy
pandas
scipy
scikit-learn
seaborn
matplotlib
tqdm

# Recommender system
scikit-surprise

# API & serving
fastapi
uvicorn

# Experimentation & Notebooks
jupyter
ipykernel

# Testing
pytest

# Kaggle API (to fetch datasets)
kaggle